Remove SQLite References and Generate Routes Index
